add Request dtos and edit Response dtos
