+++ foreignExchange/src/main/java/com/example/foreignexchange/dto/TransactionResponse.java
@@ -6,25 +6,29 @@ import java.time.LocalDateTime;
 public class TransactionResponse {
 
 
-        @Schema(description = "Transaction ID", example = "27")
-        private int id;
-
-        @Schema(description = "Source currency code", example = "EUR")
-        private String sourceCurrency;
-
-        @Schema(description = "Target currency code", example = "BGN")
-        private String targetCurrency;
-
-        @Schema(description = "Source amount", example = "100")
-        private double sourceAmount;
+//        @Schema(description = "Source currency code", example = "EUR")
+//        private String sourceCurrency;
+//
+//        @Schema(description = "Target currency code", example = "BGN")
+//        private String targetCurrency;
+//
+//        @Schema(description = "Source amount", example = "100")
+//        private double sourceAmount;
 
         @Schema(description = "Target amount", example = "195.58")
         private double targetAmount;
+    @Schema(description = "Transaction ID", example = "27")
+    private int id;
 
-        @Schema(description = "Timestamp of the transaction", example = "2024-04-07T21:58:46.9225672")
-        private LocalDateTime timestamp;
+//        @Schema(description = "Timestamp of the transaction", example = "2024-04-07T21:58:46.9225672")
+//        private LocalDateTime timestamp;
 
     public TransactionResponse() {
+    }
+
+    public TransactionResponse(int id, double targetAmount) {
+        this.targetAmount = targetAmount;
+        this.id = id;
     }
 
     public int getId() {
@@ -35,29 +39,29 @@ public class TransactionResponse {
         this.id = id;
     }
 
-    public String getSourceCurrency() {
-        return sourceCurrency;
-    }
-
-    public void setSourceCurrency(String sourceCurrency) {
-        this.sourceCurrency = sourceCurrency;
-    }
-
-    public String getTargetCurrency() {
-        return targetCurrency;
-    }
-
-    public void setTargetCurrency(String targetCurrency) {
-        this.targetCurrency = targetCurrency;
-    }
-
-    public double getSourceAmount() {
-        return sourceAmount;
-    }
-
-    public void setSourceAmount(double sourceAmount) {
-        this.sourceAmount = sourceAmount;
-    }
+//    public String getSourceCurrency() {
+//        return sourceCurrency;
+//    }
+//
+//    public void setSourceCurrency(String sourceCurrency) {
+//        this.sourceCurrency = sourceCurrency;
+//    }
+//
+//    public String getTargetCurrency() {
+//        return targetCurrency;
+//    }
+//
+//    public void setTargetCurrency(String targetCurrency) {
+//        this.targetCurrency = targetCurrency;
+//    }
+//
+//    public double getSourceAmount() {
+//        return sourceAmount;
+//    }
+//
+//    public void setSourceAmount(double sourceAmount) {
+//        this.sourceAmount = sourceAmount;
+//    }
 
     public double getTargetAmount() {
         return targetAmount;
@@ -67,11 +71,11 @@ public class TransactionResponse {
         this.targetAmount = targetAmount;
     }
 
-    public LocalDateTime getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(LocalDateTime timestamp) {
-        this.timestamp = timestamp;
-    }
+//    public LocalDateTime getTimestamp() {
+//        return timestamp;
+//    }
+//
+//    public void setTimestamp(LocalDateTime timestamp) {
+//        this.timestamp = timestamp;
+//    }
 }
